sll files

+++ match.cpp
@@ -19,35 +19,35 @@ int getMatchAnimalRank(Animal**);
 
 /*getters for the match class attributes*/
 unsigned int Match::getMatchId(){return matchId;}
-unsigned int Match::getNonPhysicalTotal(){return matchNonPhysicalTotal;}
-unsigned int Match::getNonPhysicalAverage(){return matchNonPhysicalAverage;}
-unsigned int Match::getPhysicalTotal(){return matchPhysicalTotal;}
-unsigned int Match::getPhysicalAverage(){return matchPhysicalAverage;}
-unsigned int Match::getMatchTotal(){return matchTotal;}
-unsigned int Match::getMatchAverage(){return matchAverage;}
+float Match::getNonPhysicalTotal(){return matchNonPhysicalTotal;}
+float Match::getNonPhysicalAverage(){return matchNonPhysicalAverage;}
+float Match::getPhysicalTotal(){return matchPhysicalTotal;}
+float Match::getPhysicalAverage(){return matchPhysicalAverage;}
+float Match::getMatchTotal(){return matchTotal;}
+float Match::getMatchAverage(){return matchAverage;}
 
 Client* Match::getClient(){return client;}    //returns a pointer to this match's Client
 Animal* Match::getAnimal(){return animal;}    //returns a pointer to this match's Animal
 
 /*setters for the match class attributes*/
-void Match::setNonPhysicalTotalAndAverage(unsigned int nonPhysicalTotal){
+void Match::setNonPhysicalTotalAndAverage(float nonPhysicalTotal){
     matchNonPhysicalTotal = nonPhysicalTotal;
     matchNonPhysicalAverage = matchNonPhysicalTotal / 12;
 }
 
-void Match::setNonPhysicalAverage(unsigned int nonPhysicalAverage){matchNonPhysicalAverage = nonPhysicalAverage;}
+void Match::setNonPhysicalAverage(float nonPhysicalAverage){matchNonPhysicalAverage = nonPhysicalAverage;}
 
-void Match::setPhysicalTotalAndAverage(unsigned int physicalTotal){
+void Match::setPhysicalTotalAndAverage(float physicalTotal){
     matchPhysicalTotal = physicalTotal;
     matchPhysicalAverage = matchPhysicalTotal / 3;
 
 }
 
-void Match::setPhysicalAverage(unsigned int physicalAverage){}
+void Match::setPhysicalAverage(float physicalAverage){}
 
 void Match::setMatchTotalAndAverage(){
     matchTotal = matchNonPhysicalTotal + matchPhysicalTotal;
     matchAverage = matchPhysicalAverage + matchNonPhysicalAverage;
 }
 
-void Match::setMatchAverage(unsigned int average){}
+void Match::setMatchAverage(float average){}
